Move new DataSource resolution rules into DataSourceResolution.
